+++ src/main/java/io/github/wouink/furnish/WoodenSet.java
@@ -2,14 +2,11 @@ package io.github.wouink.furnish;
 
 import io.github.wouink.furnish.block.*;
 import io.github.wouink.furnish.block.util.ShapeHelper;
-import io.github.wouink.furnish.item.CrateBlockItem;
 import io.github.wouink.furnish.reglib.RegLib;
-import net.minecraft.core.component.DataComponents;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.Identifier;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.Items;
-import net.minecraft.world.item.component.ItemContainerContents;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockBehaviour;
 import net.minecraft.world.level.block.state.properties.WoodType;
@@ -55,10 +52,8 @@ public class WoodenSet {
         ladder = RegLib.registerBlock(wood + "_ladder", Ladder::new, props.noOcclusion(), true);
         table = RegLib.registerBlock(wood + "_table", Table::new, props.noOcclusion().forceSolidOn(), true);
 
-        crate = RegLib.registerBlock(wood + "_crate", Crate::new, props, false);
+        crate = RegLib.registerBlock(wood + "_crate", Crate::new, props, true);
         FurnishContents.crates.add(crate);
-        // using custom BlockItem as we need a custom tooltip
-        RegLib.registerCustomBlockItem(crate, CrateBlockItem::new, new Item.Properties().component(DataComponents.CONTAINER, ItemContainerContents.EMPTY));
 
         shelf = RegLib.registerBlock(wood + "_shelf", Shelf::new, props.noOcclusion().strength(.7f), true);
         FurnishContents.shelves.add(shelf);