+++ src/main/java/io/github/wouink/furnish/datagen/FurnishLootTablesGenerator.java
@@ -51,7 +51,7 @@ public class FurnishLootTablesGenerator extends FabricBlockLootTableProvider {
         for(WoodenSet set : FurnishContents.WOODEN_SETS.values()) {
             for(Block b : set.getAllBlocks()) {
                 if(b == null) continue;
-                else if(b == set.crate) add(b, createShulkerBoxDrop(b)); // TODO not working
+                else if(b == set.crate) add(b, createShulkerBoxDrop(b));
                 else dropSelf(b);
             }
         }